feat(azurite): equity-bucketed blueprint selected by local exploitability
Add `cfrgen exploit`, a best-response evaluator over the abstract game
that scores a blueprint (or a bare heuristic) as BB/100 exploitability,
so blueprints can be selected locally instead of by hosted matches.

- internal/cfr: Strategy interface, Exploiter, equity-bucketed hand
  profile, layout transitions, and tests
- internal/lapis: Purify and Greedy become build flags; pluggable
  Fallback for lost hands and untrained sets
- cmd/bot: "azurite" profile samples the trained mixture and falls back
  to the onyx lines
- Makefile: cfrgen-azurite, exploit-azurite, bot-azurite(-release)
  targets that overlay the uncommitted blueprint at build time

+++ internal/cfr/refine_test.go
@@ -18,7 +18,7 @@ func TestRichFinalAbstractionRefinesLegacy(t *testing.T) {
 			continue
 		}
 		old := base.Classes[id]
-		if int(info.Final) >= rich.FinalBuckets || rich.finalParents[info.Final] != old.Final {
+		if int(info.Final) >= rich.FinalBuckets || uint16(rich.finalParents[info.Final]) != old.Final {
 			t.Fatalf("class %d has invalid refinement", id)
 		}
 		info.Final = old.Final
